ring-trace: display all unknown hops separately in the output graph
Currently, when there are several unknown hops in a traceroute, they are
all mapped to a single node when displayed.  This makes the resulting
graph really confusing.

Instead, make sure that all unknown hops are displayed separately in the
output graph.

Example graphs before this commit:
https://files.polyno.me/grenode/trace-grenode-ovh.svg
https://files.polyno.me/grenode/trace-grenode-ovh2.svg

Same graph after this commit:
https://files.polyno.me/grenode/trace-grenode-ovh-test.svg

A similar change could be done to the ASN graph (-a option), but
unfortunately it's more difficult to do right.